requirements.txt
# Core AI — Groq (FREE, replaces Anthropic)
groq>=0.9.0

# Market Data
yfinance>=1.0.0

# Data Processing
pandas==2.2.2
numpy==1.26.4

# HTTP & Web
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.1
feedparser==6.0.11

# Config
python-dotenv==1.0.1

# UI / Formatting
rich==13.7.1
tabulate==0.9.0
colorama==0.4.6

# Utils
python-dateutil==2.9.0
asyncio==3.4.3